moved makeFatJets to utils + bugfix in Filler/BambuProdAod
